+++ apps/client/src/App.tsx
@@ -1,8 +1,10 @@
-import { Welcome } from './pages/welcome/Welcome';
+import { Header } from './components/header/Header';
+import { Welcome } from './components/pages/welcome/Welcome';
 
 function App() {
   return (
-    <div className="min-h-screen flex justify-center items-center bg-[url(./src/assets/images/background-light.jpg)] bg-no-repeat bg-fixed bg-center bg-cover font-rub select-none">
+    <div className="min-h-screen flex flex-col justify-center items-center bg-[url(../public/images/background-light.jpg)] bg-no-repeat bg-fixed bg-center bg-cover font-rub select-none">
+      <Header />
       <Welcome />
     </div>
   );
