+++ apps/client/src/components/header/Header.tsx
@@ -1,0 +1,3 @@
+export function Header() {
+  return <div className="bg-amber-300 w-screen max-w-7xl p-3">Header</div>;
+}
